ext4: Add support for FIDTRIM, a best-effort ioctl for deep discard trim
* What
This provides an interface for issuing an FITRIM which uses the
secure discard instead of just a discard.
Only the eMMC command is "secure", and not how the FS uses it:
due to the fact that the FS might reassign a region somewhere else,
the original deleted data will not be affected by the "trim" which only
handles un-used regions.
So we'll just call it "deep discard", and note that this is a
"best effort" cleanup.

* Why
We want to be able to cleanup most of the unused blocks.
We don't want to constantly secure-discard via a mount option.

>From an eMMC spec perspective, it tells the device to really get rid of
all the data for the specified blocks and not just put them back into the
pool of free ones. The eMMC spec says the secure trim handling must
makes sure the data (and metadata) is not available anymore.
  JEDEC Standard No. 84-A441
  7.6.9 Secure Erase
  7.6.10 Secure Trim

Signed-off-by: Geremy Condra <gcondra@...gle.com>
Signed-off-by: JP Abgrall <jpa@...gle.com>