mm: page_alloc: introduce alloc_contig_range()
This commit adds the alloc_contig_range() function which tries
to allocate given range of pages.  It tries to migrate all
already allocated pages that fall in the range thus freeing them.
Once all pages in the range are freed they are removed from the
buddy system thus allocated for the caller to use.

Signed-off-by: Michal Nazarewicz <mina86 at mina86.com>
Signed-off-by: Marek Szyprowski <m.szyprowski at samsung.com>
Acked-by: Mel Gorman <mel at csn.ul.ie>
Tested-by: Rob Clark <rob.clark at linaro.org>
Tested-by: Ohad Ben-Cohen <ohad at wizery.com>
Tested-by: Benjamin Gaignard <benjamin.gaignard at linaro.org>