mm: trigger page reclaim in alloc_contig_range() to stabilise watermarks
alloc_contig_range() performs memory allocation so it also should keep
track on keeping the correct level of memory watermarks. This commit adds
a call to *_slowpath style reclaim to grab enough pages to make sure that
the final collection of contiguous pages from freelists will not starve
the system.

Signed-off-by: Marek Szyprowski <m.szyprowski at samsung.com>
Signed-off-by: Kyungmin Park <kyungmin.park at samsung.com>
CC: Michal Nazarewicz <mina86 at mina86.com>
Tested-by: Rob Clark <rob.clark at linaro.org>
Tested-by: Ohad Ben-Cohen <ohad at wizery.com>
Tested-by: Benjamin Gaignard <benjamin.gaignard at linaro.org>
Tested-by: Robert Nelson <robertcnelson at gmail.com>